stats: Calculate a min/max for each measurement
PRTG doesn't support min/max/median/percentile calculation for
values submitted. Just storing the average of the last report
interval is a bit too limited. Calculate min/max as well and
report it.

Build show around showOn that takes a function that can print
it. Use it in the doctest to print the output in the test.